Layla_Voicebot/requirements.txt
crewai
streamlit-audiorecorder
openai
pyttsx3
pydub
streamlit
python-dotenv
pinecone
markitdown
pydantic